Bump edu.hm.hafner:codingstyle-pom from 3.32.0 to 4.5.0

+++ src/main/java/edu/hm/hafner/echarts/line/LinesDataSet.java
@@ -153,7 +153,7 @@ public class LinesDataSet {
     }
 
     private double round(final Double value) {
-        return (double) Math.round(value * Math.pow(10, decimals)) / Math.pow(10, decimals);
+        return Math.round(value * Math.pow(10, decimals)) / Math.pow(10, decimals);
     }
 
     public List<Integer> getBuildNumbers() {
